fix: address PRRT_kwDOSJAM6s6dNBTY — validate linked gitdir before ensure no-op
Require the linked Git admin dir to exist and reciprocally register the
checkout before treating a surviving worktree path as usable. Extract
linked-worktree helpers so git_manager stays under the 1500-line gate.
